bot.{staging.}kernelci.org/nodes.yaml: comment-out disabled runners
Update the nodes.yaml files to comment out all the disabled runners.
The Azure clusters are now all disabled as well as VM runners for
bisections and rootfs images.  Only the production rootfs builder
remains defined in the file as we may still need to build some
occasionally.  The runner will need to be manually put offline in
Jenkins when the VM is not running.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>